+++ TeamNabf/NabfProject/Events/ReceivedJobEvent.cs
@@ -21,7 +21,7 @@ namespace NabfProject.Events
 
 		public override string ToString()
 		{
-            int desire;
+            double desire;
             if (Notice.TryGetDesirabilityOfAgent(Receiver, out desire))
                 return "Recieved job: " + Notice.GetType().Name + " " + Notice.Id + " desire: " + desire;
             else
